+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## vNext
 
-- Add load forecasting to NYISO, PJM, CAISO, ISONE, Ercot, SPP, MISO
+- Added load forecasting to NYISO, PJM, CAISO, ISONE, Ercot, SPP, MISO
+- Removed yesterday methods. Use `get_historical_*()` instead
 
 ## v0.6.0 - Aug 17, 2022